more work on #1 - spec2 doesn't handle splitters well
This is a "best I can do for now" fix for #1. Resizing panes in the browser proved to be problematic and even where it "works" the visual results aren't very useful. For now, there are splitters between the project/package panes & category/methods pane. The latter doesn't enlarge well at all but it might help a little.