completion engine default fix and cancel>method change fix
after cancel, methods could not be changed. We weren't telling the pane that hasUnacceptedEdits was false. This is because we now set autoAccept to false. 

The last commit was not the correct fix as changing the defautl completion engine affected Pharo windows. Instead, just return the Jadeite completion class directly.