Methods not saved - fixed recently introduced bug
In the course of making the change marker appear in the browser's source text (which involved morphic hacks) the underlying morphic editor shows correctly in the pane when typing but the spec2 presenter wasn't in sync. Therefore the method text that was saved was wrong. We force them to be synced on save. Eventually a better fix will appear, I hope. 

Also, renamed JadeiteSmalltalkEditor and made sure cancel makes the method list presenter source pane's hasUnacceptedEdits: be false.