Changes for code styling were preventing some method compilation
Changes in this commit: 
https://github.com/GemTalk/JadeiteForPharo/commit/a3297b21e906ad5ca477a688220353cac9742bcd 
were preventing instance-side keyword methods in RowanService from being able to be compiled. 

Class-side method #lookupVar: was redirecting the lookup for server code in an attempt to get the server's parse tree. 
Since those changes were put-aside, we can back out those changes.  (Technically, not backed out, but instead using the super class so as to not lose the code for future work)