End key goes to end of source code line now
Looks like Pharo wants the source line endings to be cr instead of lf to recognize where lines are. 
Not mentioned in the issue, the home key was going to the beginning of the source. Now goes to beginning of line to match end. 

https://github.com/GemTalk/JadeiteForPharo/issues/19

Added comment lines of code in #setupRSRConnection that will do server side RSR tracing.
Some changes to JadeiteMorphicCodeAdapter shows where to hook into key bindings (for future). Default to super implementation for other methods for now.