Fixes #17 - closing debugger after ad-hoc inspection
https://github.com/GemTalk/JadeiteForPharo/issues/17 
If we got an error during an ad-hoc inspection and terminated the execution, Jadeite  foolishly tried to open the inspector. Now, we look at the service's oop before trying to show the inspector window and make sure we have a valid oop. 
Cleaned up an invalid comment. 
Also, discovered that inspecting an element of an inspected object broke because we were sending an invalid `super` message. Fixed.