Fixes #30 - walkback saving method
If the hierarchy tab was selected and there was no class list selection then you'd get a walkback. #selectMethodsFrom: was trying to compare the selected class in the class list to the method's class. Should use the selected class from the open tab. 

Refactored to get one less send of `isString`

https://github.com/GemTalk/JadeiteForPharo/issues/30