Fixes #29 - git repository sha display
Added menu option to connection launcher that will show the commit id's of the appropriate JadeiteForPharo repositories that are loaded as well as the sha of RowanClientServices if connected to a server. Display is shown in a workspace that can be copy/pasted. If no connection, only client repositories sha's are shown. 

Also removed a few unsent methods. 

https://github.com/GemTalk/JadeiteForPharo/issues/29