scroll bars covering defined text + clear defined text on package/project switch
Fixed #59 - I didn't see an easy way to get hover help to show class names but middle mouse will give you the label name. This is the 3rd checkbox in the issue. 
https://github.com/GemTalk/JadeiteForPharo/issues/59

Also, when saving methods in the method list browser, ensure that the selected item's meta is passed along.