fixes for changes browser, project changes,  selecting methods
Changes browser was not showing the class name correctly if a class-side method was changed. 
When opening changes browser, don't do so if project is not on disk. Rowan will get upset. 
Moving from class to instance side and vice versa, and name of method exists in both, refresh source.