Save Server File improvements. Fixed #57.
Save file is a little better. 
The defaults when saving a file are the name and directory of the open file. 
Writing a file wasn't truncating the original file. 
Unaccepted edits wasn't being reset after save. 

Two unrelated changes - don't try to return from a block on doIt and printIt. 

https://github.com/GemTalk/JadeiteForPharo/issues/57
https://github.com/GemTalk/RowanClientServices/commit/f37572f9d408b8acab22d2b70109b449e0d5a499