work in progress - forking commands so semaphores won't block
All part of the JfP framework rewrite to handle RSR's asynchronous nature better. 
Introducing a semaphore into the basic command issuing mechanism of RowanServiceUpdater lets us treat RSR as a synchronous operation. 
However, that blocked the gui. So now we fork commands into a background thread which won't block. 

To handle triggered events, we look at the process issuing the command. If it's the main UI thread, we fork a process 
but if the process is not the main UI thread, we do not fork thus allowing triggered events which issue commands to proceed synchronously. 

May not be foolproof but it's working better so far than anything previously tried.