work in progress - fix side effects in single threaded rsr calls
updates can't be performed in the promise completion block because gui updates must happen in the Morphic ui process. 
refactor redundate updatesPerform* code
Ignore a command is issued but is already in command - these are updates that can be safely ignored. Previously, we were waiting in a critical block but that hung the ui process. 
Restore the removing of unprocessed updates. Not sure if this is correct yet, but my believe is that subsequent updates supercede prior updates. 
break complex if true code into separate methods. 
#computeVisibleMethods: - can't remember but I don't think we just want to display methods in the selected package as that's confusing. Especially since we have color coding to tell us where methods are. 
Clear filter list selections on class selection.