terminate the process that issued command if a debugger was terminated
If we debug on the server and the process is terminated, terminate the client process that made the server call. 
This seems the best way to handle a bad situation as there is no way to know what other problems may arise if bogus results are returned from any server call. 
Requires RowanClientServices update.