Disconnect sped up significantly + defer applying updates
When Jadeite is disconnected, it closes all of it's windows except the launcher. This was very slow
because the windows were found by doing subinstances of JadeitePresenter, JadeiteTextPresenter. 
subinstances in Pharo is very slow. Sped this up by having the connectionProfile hold onto any
window created and then closing them all at once which is much quicker. 

After a service is executed, changes to services are applied. These are mostly gui operations
but those updates weren't applied in the gui thread. Suspend the queue until they are applied to
make sure another service doesn't sneak in while updates are applied.