moved ui process handling back out of debugger code
Previous attempts to keep a ui process running to handle a debugger failed because proceeding a process in pharo which is waiting on a local semaphore proceeds through the semaphore. That doesn't seem correct but it is what it is. 
Now, if a command is sent through the ui process, open a new ui process before waiting on the current ui process's command. Then terminate the new ui process after the commandSemaphore is signaled by the completion  or catch block.