Fixed #92 - hierarchy doesn't show if class side selected
After creating the hierarchy, when the class was to be selected, the equality test wasn't passing during lookup because class equality is based on class meta, too. 
Since the rest of the structure (rightly or wrongly) doesn't have it's meta set, might as well do the look up based on meta false. (meta is corrected after being clicked) Think this is safe. 
FWIW - There's more work to do on the hierarchy display. 
https://github.com/GemTalk/JadeiteForPharo/issues/92