Fixed: #104 - End button not working
This helped complete the fix for earlier bug regarding copying source outside pharo and getting cr's in the text on linux. 
When you type in a pharo workspace, the Enter key will insert a cr but we use lf's to avoid the above-mentioned problem. Now end/home buttons check for the first cr or lf to cover both types of text. 
Not sure I'm super-thrilled with the mixing of end of line characters but it seems to work. 
https://github.com/GemTalk/JadeiteForPharo/issues/104

Also, fixed problem where dictionary pane would go blank if project changed. 
Removed unsent methods.