Fixed #100 - Add/Clear breakpoint enables properly now
If you right click to open the context menu, the point where you right-clicked will determine the stepPoint selected. If it's not a stepPoint, Add/Clear Breakpoint will not be enabled. 

Note to self - The magic code  (which was not easy to unearth) is in #selectedStepPoint. The secret is to get the global cursor point and convert it to a local point relative to the *textArea* ... not the widget itself.
After selecting the menu item, the mouse will have moved so the step point must be computed on first right-mouse click which happens to be done during menu item enablement. 

https://github.com/GemTalk/JadeiteForPharo/issues/100