format source shortcut fixed
Ctrl-Shift-F opens a Find Class browser in JfP but was breaking if the method source was selected. 
Also, 
Added a transcript to at least know if we still are getting arrays in a ListModel collection. 
Alt-F reformats a method source now
Don't reset test class list - don't think it was needed and was resetting selections. 
find class should always run in the UI process