Improvement for #109 - sticky step point underlines
Not sure if this is the ultimate fix for #109 but a definite improvement. After a method service replication, we were updating the browser's method source pane twice in JadeiteBrowserPresenter>>#updateMethodSource & RowanMethodService>>#methodSourceUpdate:browser:. By eliminating the latter method & registration, we not only have one less update to perform after a server round trip, it seems to draw the step point lines better, too. Still not sure why but I guess that it has to do with doing deferred UI updates in a non-Morphic process. 

Summary of changes: 
* Remove RowanMethodService>>#methodSourceUpdate:browser: and browsers that sent it. 
* Removed several unused methods. 
* Setting default method source encapsulated better & resets text decorations.
* #updateMethodSource now actually updates the method source. 
* Elimination of Lint Error by sending #ifNil: in #value method now. 

https://github.com/GemTalk/JadeiteForPharo/issues/109