service command semaphores could get mixed up
Service commands use a semaphore to ensure services wait and return to the proper pharo process. But these could get overwritten if services are submitted faster than they are completed. The inst var commandSemaphore needed to be a temporary, rather than an instance variable. 
Also, when closing a window, make sure the activePresenter is not nil. The activePresenter could be nil if there was an error opening the window such as a closed rsr connection.