remove references to service queue
The service queue became nothing but a glorified forked block once the single threading code was removed. Going to keep the methods on BlockClosure but not send them as we might need single threading in the future - say for resurrecting the gui tests. Replace them with `fork` and `defer:` as appropriate. 
Removed some logging code.