signal command semaphore earlier
With the introduction of `handleServiceCompletion` we've moved gui updates for some operations into the completion block which is earlier than before. 
I think it's reasonable to signal the command semaphore before doing these gui updates so the tools can send additional services after completing a command. 
Also, removed send of #value: in saveMethodIn: ... Need to confirm this, but don't think it's necessary.