possible fix for sticky step point indicators
I'm not sure why (a bug in spec2?) but removing text decorations may have been leaving them around permanently. 
The only other place in the code where we either add or remove text decorations, they are removed & added in the same method. Taking the removal out of #setDefaultMethod fixed the reproduction case. We shall see if this can be put to bed. 

Moved a couple methods to new protocols.