possible fix for disconnected category problem + debugger step points
There have been unreproducable problems where the category list seems to become disconnected from the method list presenter. Click a category and it doesn't update the method list. 
It appears that there is a small hole in the logic that prevents infinite recursion in Pharo's ObservableValueHolder>>value:.  Once the lock instvar is set to true, there is a small window such that if the process is terminated, the ensure block to set lock back to false won't get run and no spec updates will run. I moved the setting of the lock instvar inside the ensure block. Needs usage and testing. 

Display step point markers in the debugger. 
Also, discovered that removing decorations in the source pane, they may not be found so you'd leave step point markers lying around sometimes. Now we just throw away the whole widget segment collection to be sure everything is gone. 
Also, instead of just ignoring segment decoration errors, at least write a note to the transcript for future debugging.