add Transcript tracing to monitor processes
Debugging code added to *basicIssueCommand: methods.  (should be removed after debugging done) 
Added jadeite print methods to Process for debugging. 
Update services now uses the completionBlock if the service completion block is not present. 
Cache the method service in the postUpdateMethodSource in case another process jumps in and unselects all in the method list.