when class selected, select category tab + other changes
The variable tab isn't used as much, I don't think, so when you change classes go to the category tab for convenience. (At least I was annoyed the behavior it used to be) 
Not sure why yet, but disconnecting at the end of the test, #test_importConnection was halting the test run. For now, no need to disconnect. 
#updateMethodSource - make sure breakpoints are highlighted even when we're in an update. No need to go to the server again if we just updated.