removed leftover restart of ui process
In the days of yore, we used multiple morphic processes (incorrectly) and needed to terminate the gui thread if we were terminating a failed rsr process. No more - was causing termination of the test process in a gui test and probably other bad stuff. 
renamed inCompletionBlock to the more descriptive alwaysForked inst var in the services.