Added copy class, new tests and fix discard changes problem
If a method source was changed and the user clicked on the same method, prompting them to discard or save - pressing discard would not clean up the method source pane. 
Added copy class feature which copies all behavior and state to a new class. 
Fixed capitalization typo in Browser Implementors Of ...
Removed some old tests superceded by a newer test. 

https://github.com/GemTalk/JadeiteForPharo/issues/133