first print in workspace; save server file problem; workspace tests
Fixed #131 - first print in workspace - see JadeiteWorkspace>>selectedText. There is a bug in Spec2 that doesn't sync the widget's text and the spec widget's text. #text method has been modified to adjust but it has a side effect of undoing the selection in the workspace if the texts are different. 
removed unnecessary fork in print it command.
added workspace tests for server dialog & ad hoc actions
server file save wasn't recording the filePathString properly. 

https://github.com/GemTalk/JadeiteForPharo/issues/131