Fixed #117 - method list no longer cutoff; browser opens scaled
The default font pointSize is used to scale the browser's display when opened. Other, smaller tools are not scaled - can scale them if needed. 
Method list pane in the browser, no longer is cutoff. There was a forced width constraint in the method category box which kept it from smoothly sliding when the browser was resized. 

https://github.com/GemTalk/JadeiteForPharo/issues/117