first cut - context menu browse class/refs + add ms time to sunit results
I'm not happy with the implementation yet, but now you can select text in a method source pane and browse the class or it's references. 
The menu item names no longer do fancy stuff to try to find a selector. Instead, the menu item name is taken directly from the selected text. That needs to be improved. 

Also, record the total server time in milliseconds the selected tests take to run. 

First cut fix for https://github.com/GemTalk/JadeiteForPharo/issues/83