Browser tests working again
found defect in Pharo 12 change - collections were suddenly changed to use identiy equality which broke a number of tests. Workaround by getting the desired string out of the list instead of passing in a string. 
A more critical bug was in #updateAfterFindClass:onCompletion: - tell downstream code not to do a refresh since we just refreshed the class. This was causing a test hang in severan class rename tests. 
ListModel>>addAll:  - protect against nil collections.