Add loadJadeiteForPharo.st file
Code in loadJadeiteForPharo.st can be executed in a Pharo image to install Jadeite For Pharo. 
Will try to resolve $ROWAN_PROJECTS_HOME environment variable first. If found, will install from that directory. If not, a file navigation window will open allowing the user to pick the directory that hold the required git repository clones. 
File will replace the code in the readme.