fix unreported bug - Browse from debugger
If the browser was finding a method in a large list of methods, sometimes a timing error would occur and the wrong method would be selected. 
Needed to ensure that the method list was sorted before selecting the index. 
Also, rename sortedProjectsFrom: to sortedFrom: since it's being used for more than just projects