Fixed #151 - import multiple connections
Add the ability to import connection files from a common directory.  There are limitations currently.  All files in the selected directory will try to be imported, for example. 
This is a stopgap until a proper session/connection browser can be built. Added a couple tests. Some refactorings to the launcher, too. 

Improved menu item name in inspector. 

Test resource had an infinite loop if the console didn't open. Use a timed semaphore instead. 

https://github.com/GemTalk/JadeiteForPharo/issues/151