using detach semantics after RSR setup caused hot gems workaround
In 3.7.2, a server bug was causing hot gems. That is subsequently fixed, but JfP cannot use the recently-added detach-gci-after-rsr-setup logic until we are on 3.7.3.1. 
Since there is a gci session present now, we can no longer use the gci to get the commit ids when logged in. So use gci to get commit ids when logged out but use services once logged in. 

https://github.com/GemTalk/JadeiteForPharo/issues/242