work in progress double-click hierarchy
Better, but still not handling double click consistently. Ensuring that we wait until the command is complete after the selection is necessary. 
category/methods disappearing problem was happening because of selectItem: call but collapse/expand path causes class to be deselected. More work to do. 
Also, 
Removed unneded comments in method
Don't save lastWindowSize of browser window was collapsed. 

https://github.com/GemTalk/JadeiteForPharo/issues/233