Fixed #233 - double click hierarchy show one level
The real difficulty of this was that two events come through to the hierarchy browser - selection and activation. Because we're operating on the same service, we needed to make sure the commands & updates were synchronized and didn't collide on the server. That involved making singleThreaded operate a little differently in the updater and doing some updates on the copy of the returned service. 
https://github.com/GemTalk/JadeiteForPharo/issues/233