invalid session notification testable #243
Send a notification if a session is deemed to be invalid. Anyone, like a test, can catch it and take additional actions. 
Added test of invalid session action. 
JadeiteTestResource>>#connectThenSignal: was not drawing the console fast enough so tests would be unnecessarily delayed. 
Make a generic method in tests that finds a random rowan services class. 
Simplify slightly a couple update methods by passing in selectedItem
add a couple more selection event suspensions

https://github.com/GemTalk/JadeiteForPharo/issues/243