Fixed #251 - git commit id menu option handle login failures better
If not connected to a stone, JfP tries to use the gci to get the commit ids.  However, if commit ids menu option was selected on a newly created image before a login had been attempted, and the login through gci failed, the gui might hang. Better error handling of that case now. 
https://github.com/GemTalk/JadeiteForPharo/issues/251

Also, remove duplicate method send in find class code.