Add double-click on splitter to console #235
Per request, double click on a splitter in the project list in the console now auto-adjusts the size of the column to the left of the splitter to be the max width of the longest string. 
Removed several not-used classes from the JfD port. 
make sure console window is open before reseting configuration

https://github.com/GemTalk/JadeiteForPharo/issues/235#issuecomment-2804190677