Work in progress #245 - method compilation failures after class compile
Instance methods failed compilation now open a method list which the user can use to change methods. 
Removed a number of methods that were duplicated in a super class. 
Handle case where category is empty when compiling a method. 
Reworked the way method lists open. 
Removed JadeTranscript which was not used. 

https://github.com/GemTalk/JadeiteForPharo/issues/245