Class category selection wrong if selecting already selected category
If you had multiple selections, then selected one of the selections individually, a selection event was not triggered. JadeiteClassCategoryTreeAdapter>>#basicSelectionChanged: `difference:` was the culprit. Removed test so event could be generated. 
Renamed tree presenter class as class tree presenter & class category tree presenter behave differently. 
Added accessor for ObservableValueHolder>>lock. Here's an example of how to trace the value of `lock` as it's not easy to get to: 
	'3-' trace. 
	((self classCategoryPresenter selection class slotNamed: #selection)
		 rawRead: self classCategoryPresenter selection) lock traceCr.
Suspend selection events in `updateClassCategoriesAndSelectAllFor:` before selecting all the categories as this sends Pharo on a path that doesn't release the classCategoryPresenter ObservableValueHolder lock var. Seems like a bug in Pharo. 
Cleaned up a couple JfPwoR tests that now pass.