Fixed #323 - drag drop leaves method categories in a bad state
The category list in the filter tabs should always contain JadeiteCategory's.  But it was getting strings put in when a drag/drop of a method to a class was done. Now we ensure the categories are updated as JadeiteCategory's. 
`initializePresenterList:` was not doing anything but was sent from several places. Removed. 
Pushed test up to the hierarchy so it runs in both jfp & jfpwor (for now non-stubbed server). 

https://github.com/GemTalk/JadeiteForPharo/issues/323