postOpenBlock was getting run earlier than expected
opening a new browser tab, then finding a class was doing the find in the postOpenBlock but that was run too early so when the select page event was triggered it removed the selection. 
removed some updates of the tab label which were redundant. 
refactored some common code in the tests.