Fixed #350 - method list scrollbars disappearing no more
After updating the method list with new methods (eg after changing classes) the presenter needed to be invalidated for the scrollbars to be updated. 
My guess is that if this update took place in the morphic process we wouldn't have had the problem but we're doing everything in forked processes.