Fixes #356 - class deselected when compiling class definition
In JfP, the newly created class should be selected and if the created class is a change to an existing class, method selection is preserved. Note that method selection depends on whether the method could be recompiled after the class definition change. 
Tests added
Noticed that #selectedMethodName in two classes didn't handle nil selection so fixed that and added tests. 
Removed some bogus methods, too.