Should fix #361 and #362 + Dr Tests random test order wasn't being recreated by random seed.
When connect is pressed, we store the client lib path in a share var. This was a recent change. Previously, the share var was stored on every update to the client lib path field. 
This change created some bugs when the share var was nil. Now, a client lib path default is returned if the share var is nil. The default is the working directory. 
Added some tests with nil share var to try to catch this problem if it manifests again. Test setup refactoring. 

The random seed problem was because the test class needed a suiteClass of JadeiteTestSuite

Removed methods `sortAspect` as unsent.